updated excel filter
