+++ database/trip.py
@@ -1,7 +1,7 @@
 import sqlite3
 import trip_preference
 
-# conn=sqlite3.connect('./database/trip.db')
+# conn=sqlite3.connect('./trip.db')
 # c=conn.cursor()
 # # Create the table trip.db with appropriate columns; no need to create again
 # c.execute("""
@@ -46,7 +46,7 @@ def create_trip(id, name, category, start_date, end_date, lead_guides_needed, to
     if msg != True:
         return msg
     
-    conn=sqlite3.connect('./database/trip.db')
+    conn=sqlite3.connect('./trip.db')
     c=conn.cursor()
     c.execute("SELECT * FROM trip WHERE trip_id=?", (id,))
     if c.fetchone() is not None:
@@ -64,7 +64,7 @@ def create_trip(id, name, category, start_date, end_date, lead_guides_needed, to
 
 #questionable: how should we return an error message?
 def get_trip_by_id(id):
-    conn=sqlite3.connect('./database/trip.db')
+    conn=sqlite3.connect('./trip.db')
     c=conn.cursor()
     if not isinstance(id, int):
         return ("Error: id must be an integer")
@@ -82,7 +82,7 @@ def update_trip(id, name, category, start_date, end_date, lead_guides_needed, to
     if msg != True:
         return msg
     
-    conn=sqlite3.connect('./database/trip.db')
+    conn=sqlite3.connect('./trip.db')
     c=conn.cursor()
     
     c.execute("SELECT * FROM trip WHERE trip_id=?", (id,))
@@ -97,7 +97,7 @@ def update_trip(id, name, category, start_date, end_date, lead_guides_needed, to
     return "Success!"
 
 def get_all_trips():
-    conn=sqlite3.connect('./database/trip.db')
+    conn=sqlite3.connect('./trip.db')
     c=conn.cursor()
     c.execute("SELECT * FROM trip")
     records = c.fetchall()
@@ -105,7 +105,7 @@ def get_all_trips():
     return records
 
 def delete_trip(id):
-    conn=sqlite3.connect('./database/trip.db')
+    conn=sqlite3.connect('./trip.db')
     c=conn.cursor()
     if not isinstance(id, int):
         return ("Error: id must be an integer")
@@ -125,7 +125,7 @@ def is_trip_type(category):
     return True
 
 def delete_all_trips():
-    conn=sqlite3.connect('./database/trip.db')
+    conn=sqlite3.connect('./trip.db')
     c=conn.cursor()
     c.execute("DELETE FROM trip")
     conn.commit()
